page title; moved ReadMore to shared components
