+++ src/js/components/covid19/assistanceListing/SpendingByCFDA.jsx
@@ -15,8 +15,8 @@ import SpendingByCFDAContainer from 'containers/covid19/assistanceListing/Spendi
 import GlossaryLink from 'components/sharedComponents/GlossaryLink';
 import { scrollIntoView } from 'containers/covid19/helpers/scrollHelper';
 import Analytics from 'helpers/analytics/Analytics';
+import ReadMore from 'components/sharedComponents/ReadMore';
 import DateNote from '../DateNote';
-import ReadMore from '../ReadMore';
 
 const overviewData = [
     {
